SPT-1969: Renamed UserIdentityHandlerFunction
Renamed UserIdentityHandlerFunction to GetUserIdentityHandler, which now
follows our preferred naming strategy.
