+++ openAPI/private-api.yaml
@@ -86,9 +86,9 @@ paths:
       x-amazon-apigateway-integration:
         httpMethod: POST
         uri:
-          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${UserIdentityHandlerFunction.Arn}:live/invocations
+          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${GetUserIdentityHandler.Arn}:live/invocations
         credentials:
-          Fn::GetAtt: UserIdentityHandlerRestApiRole.Arn
+          Fn::GetAtt: GetUserIdentityHandlerRestApiRole.Arn
         passthroughBehavior: NEVER
         contentHandling: CONVERT_TO_TEXT
         type: AWS_PROXY
